changed most of no_top out of the model area because we need to retrain top. but i'll stop working on it to first get a fully functional model. Modified datasplit and tofaces as needed to work with colab better

+++ tofaces.py
@@ -17,7 +17,7 @@ def string_to_face(pixels,width,height):
 #        reshaping the array into a table with a value for each pixel
         img = img.reshape(width,height)
 #        converting value array to image
-        img = im.fromarray(img)
+        img = im.fromarray(img.astype('uint8'))
 #        adding the image to the final array of images to output
         faces.append(img)
     return faces
